requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
Django @ file:///tmp/build/80754af9/django_1617740316746/work
django-countries==7.2.1
django-crispy-forms==1.13.0
djangorestframework==3.13.1
Pillow==9.0.0
psycopg2 @ file:///C:/ci/psycopg2_1612298715048/work
python-decouple==3.5
pytz==2021.3
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
wincertstore==0.2
